# perf: improve marshal and unmarshal slice value