Sample Benchmark:
```js
goos: darwin
goarch: amd64
pkg: github.com/muktihari/fit/proto
cpu: Intel(R) Core(TM) i5-5257U CPU @ 2.70GHz
                     │ valueold.txt │             valuenew.txt             │
                     │    sec/op    │    sec/op     vs base                │
ValueMarshalAppend-4    160.8n ± 4%   100.9n ± 12%  -37.27% (p=0.000 n=10)

                     │ valueold.txt │          valuenew.txt          │
                     │     B/op     │    B/op     vs base            │
ValueMarshalAppend-4     0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal

                     │ valueold.txt │          valuenew.txt          │
                     │  allocs/op   │ allocs/op   vs base            │
ValueMarshalAppend-4     0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal
```